int to float in average counters

+++ week_05/day_1_jukebox/JukeBox.h
@@ -13,7 +13,7 @@ public:
   void addSong(Song& song);
   void getEverySong();
   void songRate(string artist_of_song, string title, unsigned int rate);
-  int getAverageSongRating(string artist_of_song);
+  float getAverageSongRating(string artist_of_song);
   float getGenreRating(string genre);
 private:
   Song** songs;
